# Explicitly ignore derivatives of argument checks