--- a/src/cholesky/lkjcholesky.jl
+++ b/src/cholesky/lkjcholesky.jl
@@ -43,8 +43,8 @@ end
 function LKJCholesky(d::Int, η::Real, _uplo::Union{Char,Symbol} = 'L'; check_args::Bool=true)
     @check_args(
         LKJCholesky,
-        (d > 0, "matrix dimension must be positive"),
-        (η > 0, "shape parameter must be positive"),
+        (d, d > 0, "matrix dimension must be positive"),
+        (η, η > 0, "shape parameter must be positive"),
     )
     logc0 = lkj_logc0(d, η)
     uplo = _char_uplo(_uplo)