--- a/src/edgeworth.jl
+++ b/src/edgeworth.jl
@@ -14,7 +14,9 @@ struct EdgeworthZ{D<:UnivariateDistribution} <: EdgeworthAbstract
     n::Float64
 
     function EdgeworthZ{D}(d::UnivariateDistribution, n::Real; check_args::Bool=true) where {D<:UnivariateDistribution}
-        check_args && @check_args(EdgeworthZ, n > zero(n))
+        ChainRulesCore.ignore_derivatives() do
+            check_args && @check_args(EdgeworthZ, n > zero(n))
+        end
         new{D}(d, n)
     end
 end
@@ -80,7 +82,9 @@ struct EdgeworthSum{D<:UnivariateDistribution} <: EdgeworthAbstract
     dist::D
     n::Float64
     function EdgeworthSum{D}(d::UnivariateDistribution, n::Real; check_args::Bool=true) where {D<:UnivariateDistribution}
-        check_args && @check_args(EdgeworthSum, n > zero(n))
+        ChainRulesCore.ignore_derivatives() do
+            check_args && @check_args(EdgeworthSum, n > zero(n))
+        end
         new{D}(d, n)
     end
 end