--- a/src/matrix/lkj.jl
+++ b/src/matrix/lkj.jl
@@ -33,9 +33,11 @@ end
 #  -----------------------------------------------------------------------------
 
 function LKJ(d::Integer, η::Real; check_args::Bool=true)
-    if check_args
-        d > 0 || throw(ArgumentError("Matrix dimension must be positive."))
-        η > 0 || throw(ArgumentError("Shape parameter must be positive."))
+    ChainRulesCore.ignore_derivatives() do
+        if check_args
+            d > 0 || throw(ArgumentError("Matrix dimension must be positive."))
+            η > 0 || throw(ArgumentError("Shape parameter must be positive."))
+        end
     end
     logc0 = lkj_logc0(d, η)
     T = Base.promote_eltype(η, logc0)
@@ -75,9 +77,11 @@ insupport(d::LKJ, R::AbstractMatrix) = isreal(R) && size(R) == size(d) && isone(
 mean(d::LKJ) = Matrix{partype(d)}(I, dim(d), dim(d))
 
 function mode(d::LKJ; check_args::Bool=true)
-    p, η = params(d)
-    if check_args
-        η > 1 || throw(ArgumentError("mode is defined only when η > 1."))
+    ChainRulesCore.ignore_derivatives() do
+        if check_args
+            _, η = params(d)
+            η > 1 || throw(ArgumentError("mode is defined only when η > 1."))
+        end
     end
     return mean(d)
 end