--- a/src/multivariate/multinomial.jl
+++ b/src/multivariate/multinomial.jl
@@ -26,21 +26,21 @@ struct Multinomial{T<:Real, TV<:AbstractVector{T}} <: DiscreteMultivariateDistri
 end
 
 function Multinomial(n::Integer, p::AbstractVector{T}; check_args::Bool=true) where {T<:Real}
-    if check_args
-        if n < 0
-            throw(ArgumentError("n must be a nonnegative integer."))
-        end
-        if !isprobvec(p)
-            throw(ArgumentError("p = $p is not a probability vector."))
+    ChainRulesCore.ignore_derivatives() do
+        if check_args
+            @check_args(Multinomial, n >= 0)
+            isprobvec(p) || throw(ArgumentError("p = $p is not a probability vector."))
         end
     end
     return Multinomial{T,typeof(p)}(n, p)
 end
 
 function Multinomial(n::Integer, k::Integer; check_args::Bool=true)
-    if check_args
-        @check_args(Multinomial, n >= 0)
-        @check_args(Multinomial, k >= 1)
+    ChainRulesCore.ignore_derivatives() do
+        if check_args
+            @check_args(Multinomial, n >= 0)
+            @check_args(Multinomial, k >= 1)
+        end
     end
     return Multinomial{Float64, Vector{Float64}}(round(Int, n), fill(1.0 / k, k))
 end