--- a/src/multivariate/multinomial.jl
+++ b/src/multivariate/multinomial.jl
@@ -28,14 +28,14 @@ end
 function Multinomial(n::Integer, p::AbstractVector{T}; check_args::Bool=true) where {T<:Real}
     @check_args(
         Multinomial,
-        n >= 0,
-        (isprobvec(p), "p is not a probability vector."),
+        (n, n >= 0),
+        (p, isprobvec(p), "p is not a probability vector."),
     )
     return Multinomial{T,typeof(p)}(n, p)
 end
 
 function Multinomial(n::Integer, k::Integer; check_args::Bool=true)
-    @check_args(Multinomial, n >= 0, k >= 1)
+    @check_args Multinomial (n, n >= 0) (k, k >= 1)
     return Multinomial{Float64, Vector{Float64}}(round(Int, n), fill(1.0 / k, k))
 end