--- a/src/univariate/continuous/beta.jl
+++ b/src/univariate/continuous/beta.jl
@@ -33,14 +33,14 @@ struct Beta{T<:Real} <: ContinuousUnivariateDistribution
 end
 
 function Beta(α::T, β::T; check_args::Bool=true) where {T<:Real}
-    @check_args(Beta, α > zero(α), β > zero(β))
+    @check_args Beta (α, α > zero(α)) (β, β > zero(β))
     return Beta{T}(α, β)
 end
 
 Beta(α::Real, β::Real; check_args::Bool=true) = Beta(promote(α, β)...; check_args=check_args)
 Beta(α::Integer, β::Integer; check_args::Bool=true) = Beta(float(α), float(β); check_args=check_args)
 function Beta(α::Real; check_args::Bool=true)
-    @check_args(Beta, α > zero(α))
+    @check_args Beta (α, α > zero(α))
     Beta(α, α; check_args=false)
 end
 Beta() = Beta{Float64}(1.0, 1.0)
@@ -69,7 +69,8 @@ function mode(d::Beta; check_args::Bool=true)
     α, β = params(d)
     @check_args(
         Beta,
-        (α > 1 && β > 1, "mode is defined only when α > 1 and β > 1."),
+        (α, α > 1, "mode is defined only when α > 1."),
+        (β, β > 1, "mode is defined only when β > 1."),
     )
     return (α - 1) / (α + β - 2)
 end