--- a/src/univariate/continuous/biweight.jl
+++ b/src/univariate/continuous/biweight.jl
@@ -8,9 +8,7 @@ struct Biweight{T<:Real} <: ContinuousUnivariateDistribution
 end
 
 function Biweight(μ::T, σ::T; check_args::Bool=true) where {T<:Real}
-    ChainRulesCore.ignore_derivatives() do
-        check_args && @check_args(Biweight, σ > zero(σ))
-    end
+    @check_args(Biweight, σ > zero(σ))
     return Biweight{T}(μ, σ)
 end