--- a/src/univariate/continuous/chi.jl
+++ b/src/univariate/continuous/chi.jl
@@ -27,9 +27,7 @@ struct Chi{T<:Real} <: ContinuousUnivariateDistribution
 end
 
 function Chi(ν::Real; check_args::Bool=true)
-    ChainRulesCore.ignore_derivatives() do
-        check_args && @check_args(Chi, ν > zero(ν))
-    end
+    @check_args(Chi, ν > zero(ν))
     return Chi{typeof(ν)}(ν)
 end
 
@@ -73,11 +71,10 @@ entropy(d::Chi{T}) where {T<:Real} = (ν = d.ν;
 
 function mode(d::Chi; check_args::Bool=true)
     ν = d.ν
-    ChainRulesCore.ignore_derivatives() do
-        if check_args
-            ν >= 1 || error("Chi distribution has no mode when ν < 1")
-        end
-    end
+    @check_args(
+        Chi,
+        (ν >= 1, "Chi distribution has no mode when ν < 1"),
+    )
     sqrt(ν - 1)
 end