--- a/src/univariate/continuous/chisq.jl
+++ b/src/univariate/continuous/chisq.jl
@@ -26,7 +26,9 @@ struct Chisq{T<:Real} <: ContinuousUnivariateDistribution
 end
 
 function Chisq(ν::Real; check_args::Bool=true)
-    check_args && @check_args(Chisq, ν > zero(ν))
+    ChainRulesCore.ignore_derivatives() do
+        check_args && @check_args(Chisq, ν > zero(ν))
+    end
     return Chisq{typeof(ν)}(ν)
 end