… Project.toml
… src/univariate/continuous/epanechnikov.jl
… src/univariate/continuous/erlang.jl
… src/univariate/continuous/exponential.jl
… src/univariate/continuous/fdist.jl
… src/univariate/continuous/frechet.jl
… src/univariate/continuous/gamma.jl
… src/univariate/continuous/generalizedpareto.jl
… src/univariate/continuous/gumbel.jl
… src/univariate/continuous/inversegamma.jl
… src/univariate/continuous/inversegaussian.jl
… src/univariate/continuous/laplace.jl
… src/univariate/continuous/levy.jl
… src/univariate/continuous/logistic.jl
… src/univariate/continuous/logitnormal.jl
… src/univariate/continuous/lognormal.jl
… src/univariate/continuous/loguniform.jl
… src/univariate/continuous/noncentralbeta.jl
… src/univariate/continuous/noncentralchisq.jl
… src/univariate/continuous/noncentralf.jl
… src/univariate/continuous/noncentralt.jl
… src/univariate/continuous/normal.jl
… src/univariate/continuous/normalcanon.jl
… src/univariate/continuous/pareto.jl
… src/univariate/continuous/pgeneralizedgaussian.jl
… src/univariate/continuous/rayleigh.jl
… src/univariate/continuous/rician.jl
… src/univariate/continuous/semicircle.jl
… src/univariate/continuous/skewedexponentialpower.jl
… src/univariate/continuous/skewnormal.jl