@sethaxen 2022-01-23
I'm not a big fan of how much ChainRules-specific code is now being included in these functions.  What if the `@check_args` macro included `ChainRulesCore.ignore_derivatives` internally? Then you would only need to call it for cases where a constructor doesn't use `@check_args`. Also, what if `@check_args` was generalized to support multiple conditions and custom error messages, which would I think allow it to be used everywhere `@check_args` currently isn't used?