@devmotion 2022-01-23
> What if the @check_args macro included ChainRulesCore.ignore_derivatives internally? Then you would only need to call it for cases where a constructor doesn't use @check_args. 

This was my initial approach but unfortunately it is not sufficient if one does not ignore the whole blocks of checks, including `if check_args`.