@devmotion 2022-01-23
> Also, what if @check_args was generalized to support multiple conditions and custom error messages, which would I think allow it to be used everywhere @check_args currently isn't used?

I considered this as well but it felt like if it becomes completely general and flexible it would become easier at some point to not use a macro but just implement the checks explicitly. It's also not only used in constructors and IIRC there's at least one case where some variables are unpacked/defined in the `check_args` block. And it felt a bit weird to include `if check_args` in the macro which would be necessary as well.

So basically my impressoon was that the macro should be able to generate something like
```julia
ChainRulesCore.ignore_derivatives() do
    if some_var # or only if check_args
        some_expr
        cond1 || throw(ArgumentError(cond1_msg))
        cond2 || ...
    end
end
```
and it was not immediately clear if and what a short but descriptive syntax could look that would provide a major advantage over implementing the block direv