@mschauer 2022-01-23
> I'm not a big fan of how much ChainRules-specific code is now being included in these functions.

This looks like fixing a problem in the wrong place.
This would now work for AD, but every _other_  nonstandard evaluation besides AD (particles, uncertainty intervals, things we don’t anticipate)  has to deal with ChainRules specific code.