@devmotion 2022-01-23
The benchmarks also show that there is no overhead for non-AD use.