@sethaxen 2022-01-23
@devmotion I can't think of a nice syntax for the macro, and your points are all valid.

I guess another approach would be to define a `_maybe_check_args(::Type{<:Distribution}, check::Bool, ps...)` function where all checks are performed for each distribution (each with their own overload). Then mark it globally as `ChainRulesCore.@non_differentiable`. But I suspect that would only work for ChainRules, since for operator-overloading ADs, one would need to dispatch on the param values, so there would be unresolved ambiguity.