@devmotion 2022-01-23
I thought about moving the checks to a separate function but this seemed unnecessarily complicated - for `ignore_derivatives` one already defines separate distribution specific functions anyway, without adding more methods to the module and hence without potential ambiguity issues.

To summarize, I think it's correct to address these performance issues in Distributions but it would be good to do it in a more developer friendly and "less special" way. I don't think it's an AD backend issue since I don't think it can be expected in general that AD neglects these checks - without us telling it that it should.

I thought about a more convenient macro. Maybe we could use something like `@check_args(Beta, a > zero(a), (b > zero(b)) => "b should be positive!") `. This would not include any additional expressions but maybe that's fine for now (even though it's a bit annoying that AD won't ignore them if they are outside of the function...).