@devmotion 2022-01-24
I modified the `@check_args` macro, inspired by `ChainRulesCore.@scalar_rule`:
````julia
    @check_args(
        D,
        @setup(statements...),
        (cond₁, message₁),
        (cond₂, message₂),
        ...,
    )

A convenience macro that generates AD-compatible checks of arguments for a distribution of
type `D`.

More concretely, it generates the following Julia code:
```julia
ChainRulesCore.ignore_derivatives() do
    if check_args
        \$(statements...)
        cond₁ || throw(ArgumentError(\$(string(D, ": ", message₁))))
        cond₂ || throw(ArgumentError(\$(string(D, ": ", message₂))))
        ...
    end
end
```

The `@setup` argument can be elided if no setup code is needed. Moreover, error messages
can be omitted. In this case the message `"the condition \$(cond) is not satisfied."` is
used.
````

I wonder though if it is too surprising that one has to define a boolean variable `check_args`. Maybe its name should be passed in the macro call, i.e., something like `@check_args(check_args, D, ...)`? On the other hand, this would feel a bit redundant.