@mschauer 2022-01-24
Why not just having a dedicated function
```julia
check_arguments(checkargs, f::Function) = checkargs && f()
ChainRulesCore.@non_differentiable check_arguments(checkargs, f)
```
This is positively saying what happens ("checking arguments") instead of negatively "can't differentiate the following code").