@devmotion 2022-01-24
Sure, this could be done but it seems equivalent to what `ignore_derivatives` does. In the current state of the PR `ignore_derivatives` only shows up in the code of the macro - and similarly the call to a `check_args` function would only show up there if we don't want to remove the macro and implement all checks explicitly.