@mschauer 2022-01-24
We would own `check_arguments ` so on the path of standard evaluation we would not call chainrules code at all and we are free to add any methods