@mschauer 2022-01-24
We can thus depend on a hypothetical `UncertaintyCore.jl` and define 
```
UncertaintyCore.@deterministic check_arguments(checkargs, f)
```
if that is required.